Shift all classes to respective packages
