--- src/main/java/JinjjaException.java
@@ -1,5 +1,0 @@
-class MissingParameterException extends RuntimeException {
-    public MissingParameterException(String message) {
-        super(message);
-    }
-}
