+++ src/main/java/jinjja/command/ExitCommand.java
@@ -1,7 +1,13 @@
+package jinjja.command;
+
+import jinjja.storage.Storage;
+import jinjja.task.TaskList;
+import jinjja.ui.Ui;
+
 /**
  * Command to exit the application.
  */
-class ExitCommand extends Command {
+public class ExitCommand extends Command {
 
     @Override
     public void execute(TaskList tasks, Storage storage, Ui ui) {
