+++ src/main/java/jinjja/exception/MissingParameterException.java
@@ -1,0 +1,10 @@
+package jinjja.exception;
+
+/**
+ * Exception thrown when a required parameter is missing.
+ */
+public class MissingParameterException extends RuntimeException {
+    public MissingParameterException(String message) {
+        super(message);
+    }
+}
