+++ src/main/java/jinjja/parser/CommandType.java
@@ -1,16 +1,10 @@
+package jinjja.parser;
+
 /**
  * Enum representing the different types of commands that the application can handle.
  */
-enum CommandType {
-    BYE,
-    LIST,
-    MARK,
-    UNMARK,
-    TODO,
-    DEADLINE,
-    EVENT,
-    DELETE,
-    UNKNOWN;
+public enum CommandType {
+    BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, UNKNOWN;
 
     /**
      * Converts a string input to the corresponding CommandType enum.
@@ -22,7 +16,7 @@ enum CommandType {
         if (input == null || input.trim().isEmpty()) {
             return UNKNOWN;
         }
-        
+
         try {
             return CommandType.valueOf(input.toUpperCase());
         } catch (IllegalArgumentException e) {
