+++ src/main/java/jinjja/task/Todo.java
@@ -1,4 +1,9 @@
-class Todo extends Task {
+package jinjja.task;
+
+/**
+ * Represents a todo task without any date/time attached to it.
+ */
+public class Todo extends Task {
     public Todo(String description) {
         super(description);
     }
